addressing change requests
