+++ .setup/install_submitty/install_bin.sh
@@ -109,7 +109,7 @@ chmod 550 ${SUBMITTY_INSTALL_DIR}/sbin/docker_cleanup.sh
 chmod 555 ${SUBMITTY_INSTALL_DIR}/sbin/killall.py
 
 # DAEMON_USER only things in sbin
-array=( auto_rainbow_grades.py auto_rainbow_scheduler.py build_config_upload.py send_email.py cleanup_old_email.py cleanup_active_graders.py send_notification.py generate_grade_summaries.py submitty_daemon_jobs run_grading_clustering.py database_queries.py clustering_algorithms)
+array=( auto_rainbow_grades.py auto_rainbow_scheduler.py build_config_upload.py send_email.py cleanup_old_email.py cleanup_active_graders.py send_notification.py generate_grade_summaries.py submitty_daemon_jobs grading_clustering)
 for i in "${array[@]}"; do
     chown -R root:"${DAEMON_GROUP}" ${SUBMITTY_INSTALL_DIR}/sbin/${i}
     chmod -R 750 ${SUBMITTY_INSTALL_DIR}/sbin/${i}
